BankMoneyTransfer/requirements.txt
mysql-connector-python
tk
Pillow